rename file to lighthouse since only using that
